move ws to core module
